Incorporate Nick's feedback. Thank you!
